+++ R/zzz.R
@@ -6,3 +6,14 @@
 #' @author Phillip B. Nicol <philnicol740@gmail.com>
 #' @name MorphoGAM
 "_PACKAGE"
+
+utils::globalVariables(c(
+  "Var1",
+  "color",
+  "fx",
+  "name",
+  "r",
+  "value",
+  "x",
+  "y"
+))